+++ Source/Tools/AngelScriptBaker.cpp
@@ -102,7 +102,7 @@ void AngelScriptBaker::BakeFiles(FileCollection files)
         }));
     }
 
-    int32 errors = 0;
+    size_t errors = 0;
 
     for (auto& file_baking : file_bakings) {
         try {
